# backend/openapi.yaml
openapi: 3.0.3
info:
  title: VeloSync API
  version: 0.1.0
  description: Accountancy Management System API
servers:
  - url: http://localhost:4000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/signup:
    post:
      summary: Sign up a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                fullName:
                  type: string
      responses:
        '200':
          description: User created successfully
        '400':
          description: User already exists
  /auth/login:
    post:
      summary: Login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
...